requirements.txt
asgiref==3.7.2
Django==4.2.2
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.18.11
gettex==1.0
Pillow==9.5.0
psycopg2==2.9.6
pylatexenc==2.10
python-decouple==3.8
python-gettext==5.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
